{{cookiecutter.project_name}}/{{cookiecutter.project_name}}-backend/requirements.txt
boto3==1.26.50
Django>=3.2,<4
django-cors-headers==3.13.0
djangorestframework==3.14.0
lxml==4.9.3
psycopg2-binary==2.9.5
requests==2.28.2
